Run a guessing (game) from the stdlib
Any parenthesis that are used in parameters will now be correctly removed
allowing you to print or return multiple words.

Add a game and several helper functions to the standard library. To play, simply
guess a number 0-100 and see what happens.

Officially removed escaped strings. I can't think of a good way to handle them
yet.

Removed (let) because it was useless.